requirements.txt
Django==2.1.8
django-extra-fields==1.1.0
django-filter==2.1.0
django-pyodbc-azure==2.1.0.0
django-rest-framework==0.1.0
djangorestframework==3.9.2
djangorestframework-queryfields==1.0.0
Pillow==6.0.0
pyodbc==4.0.26
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
virtualenv==16.4.3
virtualenvwrapper-win==1.2.5
